back/requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.1.2
Automat==22.10.0
backcall==0.2.0
backports.zoneinfo==0.2.1
blinker==1.8.2
Brotli==1.0.9
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==2.0.4
click==8.1.7
cmake==3.21.0
colorama==0.4.6
constantly==23.10.4
contourpy==1.2.1
cryptography==42.0.8
cycler==0.11.0
daphne==4.1.2
decorator==5.1.1
dill==0.3.8
distro==1.9.0
Django==4.2.13
django-cors-headers==4.4.0
django-crontab==0.7.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dlib==19.24.0
docker-pycreds==0.4.0
exceptiongroup==1.2.1
executing==2.0.1
faiss-cpu==1.8.0
ffmpeg-python==0.2.0
filelock==3.13.1
Flask==3.0.3
Flask-Cors==4.0.1
flatbuffers==24.3.25
fonttools==4.51.0
fsspec==2024.5.0
future==1.0.0
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.43
gmpy2==2.1.2
google-auth==2.29.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.64.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
hyperlink==21.0.0
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
incremental==22.10.0
intel-openmp==2021.4.0
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
keras==2.13.1
kiwisolver==1.4.4
libclang==18.1.1
lmdb==1.4.1
Markdown==3.6
MarkupSafe==2.1.3
matplotlib==3.7.5
matplotlib-inline==0.1.7
mkl==2021.4.0
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
mpmath==1.3.0
msgpack==1.0.8
mtcnn==0.1.1
Naked==0.1.32
networkx==3.1
ninja==1.11.1.1
numpy==1.24.3
oauthlib==3.2.2
openai==1.35.7
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
opt-einsum==3.3.0
outcome==1.3.0.post0
packaging==24.0
pandas==2.0.3
parso==0.8.4
patch-ng==1.17.4
pickleshare==0.7.5
pillow==10.3.0
pip==24.0
platformdirs==4.2.2
prompt_toolkit==3.0.46
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.8.0
pydantic_core==2.20.0
Pygments==2.18.0
PyJWT==2.9.0
pyOpenSSL==24.1.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
redis==5.0.7
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
selenium==4.23.1
sentry-sdk==2.2.0
service-identity==24.1.0
setproctitle==1.3.3
setuptools==70.3.0
shellescape==3.8.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
sparse-dot-topn==1.1.1
sqlparse==0.5.0
stack-data==0.6.3
sympy==1.12
tbb==2021.12.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
tqdm==4.66.4
traitlets==5.14.3
trio==0.26.0
trio-websocket==0.11.1
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.35
ultralytics-thop==2.0.0
urllib3==2.2.1
wandb==0.17.0
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.0.3
wget==3.2
wheel==0.43.0
win-inet-pton==1.1.0
wrapt==1.16.0
wsproto==1.2.0
zipp==3.19.1
zope.interface==6.4.post2
